+++ src/main/java/com/twilio/rest/sync/v1/service/document/DocumentPermission.java
@@ -83,7 +83,7 @@ public class DocumentPermission extends Resource {
     /**
      * Create a DocumentPermissionUpdater to execute update.
      * 
-     * @param pathServiceSid Sync Service Instance SID.
+     * @param pathServiceSid Sync Service Instance SID or unique name.
      * @param pathDocumentSid Sync Document SID or unique name.
      * @param pathIdentity Identity of the user to whom the Sync Document
      *                     Permission applies.