+++ src/main/java/com/twilio/rest/sync/v1/service/synclist/SyncListPermissionUpdater.java
@@ -32,7 +32,7 @@ public class SyncListPermissionUpdater extends Updater<SyncListPermission> {
     /**
      * Construct a new SyncListPermissionUpdater.
      * 
-     * @param pathServiceSid Sync Service Instance SID.
+     * @param pathServiceSid Sync Service Instance SID or unique name.
      * @param pathListSid Sync List SID or unique name.
      * @param pathIdentity Identity of the user to whom the Sync List Permission
      *                     applies.