+++ src/main/java/com/twilio/rest/sync/v1/service/syncmap/SyncMapPermission.java
@@ -83,7 +83,7 @@ public class SyncMapPermission extends Resource {
     /**
      * Create a SyncMapPermissionUpdater to execute update.
      * 
-     * @param pathServiceSid Sync Service Instance SID.
+     * @param pathServiceSid Sync Service Instance SID or unique name.
      * @param pathMapSid Sync Map SID or unique name.
      * @param pathIdentity Identity of the user to whom the Sync Map Permission
      *                     applies.