+++ src/main/java/com/twilio/rest/sync/v1/service/syncmap/SyncMapPermissionUpdater.java
@@ -32,7 +32,7 @@ public class SyncMapPermissionUpdater extends Updater<SyncMapPermission> {
     /**
      * Construct a new SyncMapPermissionUpdater.
      * 
-     * @param pathServiceSid Sync Service Instance SID.
+     * @param pathServiceSid Sync Service Instance SID or unique name.
      * @param pathMapSid Sync Map SID or unique name.
      * @param pathIdentity Identity of the user to whom the Sync Map Permission
      *                     applies.